LPD-30357 - Refactor: Move impl from web to service and rename to importer/exporter to match what the classes actually do.
# breaking

## What /modules/apps/style-book/style-book-api/bnd.bnd

Update version

## Why

Due to refactoring, some classes were renamed and moved to a proper module

----
