+++ modules/apps/style-book/style-book-api/src/main/resources/com/liferay/style/book/importer/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
