--- modules/apps/style-book/style-book-api/src/main/resources/com/liferay/style/book/zip/processor/packageinfo
@@ -1,1 +1,0 @@
-version 2.0.0
